+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 Changelog
-===
+=========
 
 Versions follows PEP440 version scheme.
 ---
@@ -12,6 +12,7 @@ Fixed
 * admin permissions are required to access Badge API
 * View on site button in the admin panel on the StatusBadge edit page opens correct URL for image
 * reduced height of the select dropdown 'Event type' in the 'Edit rules' popup
+* fix: [RGG-595] an admin can delete status
 
 gamma-core 3.2 (2021-10-18)
 ---